Discord.js v12 update

+++ src/commands/say.js
@@ -1,16 +1,20 @@
 module.exports = {
 	name: 'say',
 	description: 'Makes the bot say something',
-	guildOnly: false,
+	guildOnly: true,
 	// aliases: [],
 	usage: '!say [something]',
 	cooldown: 5,
 
 	execute(message, args) {
+		if (!args.length) {
+			return message.channel.send('Say what?');
+		}
+
 		const sayMessage = args.join(' ');
 		message.delete().catch(error=>{
 			console.log(error);
 		});
-		message.channel.send(sayMessage);
+		return message.channel.send(sayMessage);
 	},
 };
